Moved the code around, added tags and layer setting
