+++ apm-sniffer/apm-sdk-plugin/micrometer-1.10.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/micrometer/MicrometerReceiverTracingHandlerInterceptor.java
@@ -27,10 +27,11 @@ import org.apache.skywalking.apm.agent.core.context.ContextCarrier;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
 import org.apache.skywalking.apm.agent.core.context.tag.Tags;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
+import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.meter.micrometer.observation.SkywalkingReceiverTracingHandler;
+import org.apache.skywalking.apm.toolkit.micrometer.observation.SkywalkingReceiverTracingHandler;
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 import org.apache.skywalking.apm.util.StringUtil;
 
@@ -39,6 +40,14 @@ import org.apache.skywalking.apm.util.StringUtil;
  * {@link SkywalkingReceiverTracingHandler}.
  */
 public class MicrometerReceiverTracingHandlerInterceptor implements InstanceMethodsAroundInterceptor {
+
+    private static final String DB_TAG_PREFIX = "jdbc";
+
+    private static final String HTTP_TAG_PREFIX = "http";
+
+    private static final String RPC_TAG_PREFIX = "rpc";
+
+    private static final String MESSAGING_TAG_PREFIX = "rpc";
 
     @Override
     public void beforeMethod(EnhancedInstance objInst, Method method, Object[] allArguments, Class<?>[] argumentsTypes,
@@ -54,13 +63,10 @@ public class MicrometerReceiverTracingHandlerInterceptor implements InstanceMeth
             }
             AbstractSpan span = ContextManager.createEntrySpan(context.getName(), contextCarrier);
             span.setComponent(ComponentsDefine.MICROMETER);
-            //            TODO: We can't really set these because ReceiverContext is super generic. We don't know what the protocol will be. We can guess it from the tags maybe? But there's no certainty that same tags will be used everywhere. OTOH
-            //            entrySpan.setLayer(SpanLayer.HTTP);
-            //            Tags.URL.set(entrySpan, httpRequest.path());
-            //            Tags.HTTP.METHOD.set(entrySpan, httpRequest.method().name());
 
         } else if ("onStop".equals(methodName)) {
             ReceiverContext<Object> context = (ReceiverContext<Object>) allArguments[0];
+            SpanLayer spanLayer = TaggingHelper.toLayer(context.getAllKeyValues());
             AbstractSpan abstractSpan = ContextManager.activeSpan();
             abstractSpan
                 .setPeer(tryToGetPeer(context))
@@ -68,6 +74,9 @@ public class MicrometerReceiverTracingHandlerInterceptor implements InstanceMeth
                     context.getContextualName()) ? context.getName() : context.getContextualName());
             context.getAllKeyValues()
                    .forEach(keyValue -> abstractSpan.tag(Tags.ofKey(keyValue.getKey()), keyValue.getValue()));
+            if (spanLayer != null) {
+                abstractSpan.setLayer(spanLayer);
+            }
             ContextManager.stopSpan();
         } else if ("onError".equals(methodName)) {
             Observation.Context context = (Observation.Context) allArguments[0];
@@ -79,14 +88,9 @@ public class MicrometerReceiverTracingHandlerInterceptor implements InstanceMeth
         if (context.getRemoteServiceAddress() != null) {
             return context.getRemoteServiceAddress();
         }
-        KeyValue uri = context.getLowCardinalityKeyValue("uri");
-        if (uri != null) {
-            return uri.getValue();
-        }
         return context.getAllKeyValues()
                       .stream()
-                      .filter(keyValue -> "uri".equalsIgnoreCase(keyValue.getKey()) || "http.url".equalsIgnoreCase(
-                          keyValue.getKey()))
+                      .filter(keyValue -> "http.url".equalsIgnoreCase(keyValue.getKey()) || "uri".equalsIgnoreCase(keyValue.getKey()))
                       .findFirst()
                       .map(KeyValue::getValue)
                       .orElse("unknown");
