+++ apm-sniffer/apm-sdk-plugin/micrometer-1.10.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/micrometer/MicrometerSenderTracingHandlerInterceptor.java
@@ -27,10 +27,11 @@ import org.apache.skywalking.apm.agent.core.context.ContextCarrier;
 import org.apache.skywalking.apm.agent.core.context.ContextManager;
 import org.apache.skywalking.apm.agent.core.context.tag.Tags;
 import org.apache.skywalking.apm.agent.core.context.trace.AbstractSpan;
+import org.apache.skywalking.apm.agent.core.context.trace.SpanLayer;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceMethodsAroundInterceptor;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.MethodInterceptResult;
-import org.apache.skywalking.apm.meter.micrometer.observation.SkywalkingSenderTracingHandler;
+import org.apache.skywalking.apm.toolkit.micrometer.observation.SkywalkingSenderTracingHandler;
 import org.apache.skywalking.apm.network.trace.component.ComponentsDefine;
 import org.apache.skywalking.apm.util.StringUtil;
 
@@ -57,6 +58,7 @@ public class MicrometerSenderTracingHandlerInterceptor implements InstanceMethod
             span.setComponent(ComponentsDefine.MICROMETER);
         } else if ("onStop".equals(methodName)) {
             SenderContext<Object> context = (SenderContext<Object>) allArguments[0];
+            SpanLayer spanLayer = TaggingHelper.toLayer(context.getAllKeyValues());
             AbstractSpan abstractSpan = ContextManager.activeSpan();
             abstractSpan
                 .setPeer(tryToGetPeer(context))
@@ -64,6 +66,9 @@ public class MicrometerSenderTracingHandlerInterceptor implements InstanceMethod
                     context.getContextualName()) ? context.getName() : context.getContextualName());
             context.getAllKeyValues()
                    .forEach(keyValue -> abstractSpan.tag(Tags.ofKey(keyValue.getKey()), keyValue.getValue()));
+            if (spanLayer != null) {
+                abstractSpan.setLayer(spanLayer);
+            }
             ContextManager.stopSpan();
         } else if ("onError".equals(methodName)) {
             Observation.Context context = (Observation.Context) allArguments[0];
